requirements.txt
pelican
mMarkdown
Pygments
PyYAML
pypdf
python-pptx
